[grug] Move the gate/up interleave helpers into _moe/common
The ragged all-to-all transport and the QuACK expert MLP both interleave the
gate and up halves of the fused w13 weight. The helpers lived in sonic_cute.py,
so the transport imported a kernel module for two array operations. Move
_interleave_gate_up and _swiglu_gate_up_backward into _moe/common.py. The
computation does not change.
